refactor(rate): move to script setup
