+++ src/packages/__VUE/rate/doc.taro.md
@@ -73,6 +73,17 @@ app.use(Rate);
 | --- | --- | --- |
 | change | 当前分值修改时时触发的事件 | 当前值 |
 
+### 类型定义 version
+
+组件导出以下类型定义：
+
+```js
+import type {
+  RateProps,
+  RateInstance
+} from '@nutui/nutui-taro';
+```
+
 ## 主题定制
 
 ### 样式变量
