+++ src/packages/__VUE/rate/index.ts
@@ -1,0 +1,11 @@
+import Rate from './rate.vue';
+import type { ComponentPublicInstance } from 'vue';
+import { withInstall } from '@/packages/utils';
+
+withInstall(Rate);
+
+export type { RateProps } from './rate.vue';
+
+export type RateInstance = ComponentPublicInstance & InstanceType<typeof Rate>;
+
+export { Rate, Rate as default };
